chore : CategorySuggester 에서 CategoryChecker 기능 분리
