+++ src/main/java/menu/businessLogic/CategorySuggester.java
@@ -8,34 +8,16 @@ import java.util.List;
 
 public class CategorySuggester {
 
+    private final CategoryChecker checker = new CategoryChecker();
+
     public List<Categories> categorySuggest(){
         List<Categories> suggestedCategory = new ArrayList<>();
         for(int categoryIndex = 0; categoryIndex < 5;){
             Categories category = Categories.values()[Randoms.pickNumberInRange(1,5)-1];
-            if(checkAndAddSuggestedCategory(suggestedCategory,category)){
+            if(checker.checkAndAddSuggestedCategory(suggestedCategory,category)){
                 categoryIndex++;
-            };
-        }
-        return suggestedCategory;
-    }
-    private boolean checkAndAddSuggestedCategory(List<Categories> suggestedCategory, Categories category){
-        if(suggestedCategory.size() == 0){
-            suggestedCategory.add(category);
-            return true;
-        }
-        if(checkThatCategoryCount(suggestedCategory,category) < 2){
-            suggestedCategory.add(category);
-            return true;
-        }
-        return false;
-    }
-    private int checkThatCategoryCount(List<Categories> suggestedCategory, Categories category){
-        int count = 0;
-        for (Categories existedCategory : suggestedCategory) {
-            if (existedCategory.equals(category)) {
-                count++;
             }
         }
-        return count;
+        return suggestedCategory;
     }
 }
